add memory pool stress test.

+++ inc/MemoryPool.hpp
@@ -13,7 +13,7 @@
 class MemoryPool
 {
   public:
-    virtual void *allocate(uint32 size) = 0;
+    virtual void *allocate(uint64 size) = 0;
     virtual int free(void *ptr) = 0;
     virtual int integrityCheck() const = 0;
     virtual void dumpToFile(const std::string& fileName, const uint32 itemsPerLine, const uint32 format) const = 0;
@@ -40,8 +40,8 @@ class MemoryPool
 
     virtual ~MemoryPool(){}
 
-    uint32      m_totalPoolSize;
-    uint32      m_freePoolSize;
+    uint64      m_totalPoolSize;
+    uint64      m_freePoolSize;
 
     // Bitfield
     unsigned    m_trashOnCreation : 1;
